# frontend/src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Food Recommender API",
...
        "contact": {
            "name": "Food Recommender Team",
            "email": "support@foodrecommender.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development server"
        },
        {
            "url": "https://api.foodrecommender.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/restaurants/google_maps/restaurant/{restaurant_id}/reviews": {
            "get": {
                "tags": [
                    "restaurants",
                    "google_maps"
                ],
                "summary": "Get Restaurant Reviews",
                "description": "Get Google Maps reviews for a restaurant",
                "operationId": "get_restaurant_reviews_restaurants_google_maps_restaurant__restaurant_id__reviews_get",
                "parameters": [
                    {
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Restaurant Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RestaurantReviews"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurants/google_maps/restaurant/{restaurant_id}/item/{item_name}/reviews": {
            "get": {
                "tags": [
                    "restaurants",
                    "google_maps"
                ],
                "summary": "Get Menu Item Reviews",
                "description": "Get reviews mentioning a specific menu item",
                "operationId": "get_menu_item_reviews_restaurants_google_maps_restaurant__restaurant_id__item__item_name__reviews_get",
                "parameters": [
                    {
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Restaurant Id"
                        }
                    },
                    {
                        "name": "item_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Item Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MenuItemReviews"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurants/google_maps/restaurant/{restaurant_id}/place_details": {
            "get": {
                "tags": [
                    "restaurants",
                    "google_maps"
                ],
                "summary": "Get Place Details",
                "description": "Get detailed place information from Google Maps",
                "operationId": "get_place_details_restaurants_google_maps_restaurant__restaurant_id__place_details_get",
                "parameters": [
                    {
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Restaurant Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaceDetails"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurants/beli/restaurant/{restaurant_id}/top_items": {
            "get": {
                "tags": [
                    "restaurants",
                    "beli"
                ],
                "summary": "Get Restaurant Top Items",
                "description": "Get top recommended items at a restaurant from Beli",
                "operationId": "get_restaurant_top_items_restaurants_beli_restaurant__restaurant_id__top_items_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "restaurant_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Restaurant Id"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BeliRestaurantTopItems"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurants/beli/user/{user_id}/profile": {
            "get": {
                "tags": [
                    "restaurants",
                    "beli"
                ],
                "summary": "Get User Profile",
                "description": "Get a user's Beli profile with restaurant reviews",
                "operationId": "get_user_profile_restaurants_beli_user__user_id__profile_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BeliUserProfile"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BeliRestaurantTopItems": {
                "properties": {
                    "restaurant_id": {
                        "type": "string",
                        "title": "Restaurant Id"
                    },
                    "restaurant_name": {
                        "type": "string",
                        "title": "Restaurant Name"
                    },
                    "top_items": {
                        "items": {
                            "$ref": "#/components/schemas/BeliTopItem"
                        },
                        "type": "array",
                        "title": "Top Items"
                    },
                    "last_updated": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Last Updated"
                    }
                },
                "type": "object",
                "required": [
                    "restaurant_id",
                    "restaurant_name",
                    "top_items",
                    "last_updated"
                ],
                "title": "BeliRestaurantTopItems"
            },
...
            "BeliUserProfile": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "restaurant_reviews": {
                        "items": {
                            "$ref": "#/components/schemas/BeliUserRestaurantReview"
                        },
                        "type": "array",
                        "title": "Restaurant Reviews"
                    },
                    "total_reviews": {
                        "type": "integer",
                        "title": "Total Reviews"
                    },
                    "last_updated": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Last Updated"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "username",
                    "restaurant_reviews",
                    "total_reviews",
                    "last_updated"
                ],
                "title": "BeliUserProfile"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MenuItemReviews": {
                "properties": {
                    "item_name": {
                        "type": "string",
                        "title": "Item Name"
                    },
                    "restaurant_id": {
                        "type": "string",
                        "title": "Restaurant Id"
                    },
                    "reviews": {
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        },
                        "type": "array",
                        "title": "Reviews"
                    },
                    "average_rating": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Average Rating"
                    },
                    "mention_count": {
                        "type": "integer",
                        "title": "Mention Count"
                    }
                },
                "type": "object",
                "required": [
                    "item_name",
                    "restaurant_id",
                    "reviews",
                    "mention_count"
                ],
                "title": "MenuItemReviews"
            },
...
            "PlaceDetails": {
                "properties": {
                    "place_id": {
                        "type": "string",
                        "title": "Place Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "formatted_address": {
                        "type": "string",
                        "title": "Formatted Address"
                    },
                    "formatted_phone_number": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Formatted Phone Number"
                    },
                    "website": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Website"
                    },
                    "rating": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Rating"
                    },
                    "user_ratings_total": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Ratings Total"
                    },
                    "price_level": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Price Level"
                    },
                    "opening_hours": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Opening Hours"
                    },
                    "photos": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Photos",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "place_id",
                    "name",
                    "formatted_address"
                ],
                "title": "PlaceDetails"
            },
...
            "RestaurantReviews": {
                "properties": {
                    "restaurant_id": {
                        "type": "string",
                        "title": "Restaurant Id"
                    },
                    "place_id": {
                        "type": "string",
                        "title": "Place Id"
                    },
                    "restaurant_name": {
                        "type": "string",
                        "title": "Restaurant Name"
                    },
                    "overall_rating": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Overall Rating"
                    },
                    "total_reviews": {
                        "type": "integer",
                        "title": "Total Reviews"
                    },
                    "reviews": {
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        },
                        "type": "array",
                        "title": "Reviews"
                    },
                    "last_updated": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Last Updated"
                    }
                },
                "type": "object",
                "required": [
                    "restaurant_id",
                    "place_id",
                    "restaurant_name",
                    "total_reviews",
                    "reviews",
                    "last_updated"
                ],
                "title": "RestaurantReviews"
            },
            "Review": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "author_name": {
                        "type": "string",
                        "title": "Author Name"
                    },
                    "author_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Author Url"
                    },
                    "language": {
                        "type": "string",
                        "title": "Language"
                    },
                    "profile_photo_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Profile Photo Url"
                    },
                    "rating": {
                        "type": "integer",
                        "title": "Rating"
                    },
                    "relative_time_description": {
                        "type": "string",
                        "title": "Relative Time Description"
                    },
                    "text": {
                        "type": "string",
                        "title": "Text"
                    },
                    "time": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Time"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "author_name",
                    "language",
                    "rating",
                    "relative_time_description",
                    "text",
                    "time"
                ],
                "title": "Review"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}